requirements.txt
flask
flask-cors
werkzeug
Pillow
PyPDF2
python-docx
numpy
opencv-python
